Confirm Dialog

+++ src/app/app.component.ts
@@ -1,13 +1,14 @@
-import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { Fattura } from './domain/fattura';
 import { FatturaDialogComponent } from './fattura-dialog/fattura-dialog.component';
 import { FattureService } from './services/fatture.service';
+import { ConfirmDialogComponent } from './confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+  styleUrls: ['./app.component.scss']
 })
 export class AppComponent implements OnInit {
   title = 'fatture-manager';
@@ -30,21 +31,23 @@ export class AppComponent implements OnInit {
     });
 
     dialogRef.afterClosed().subscribe(result => {
-      const fattura: Fattura = {
-        nomeServizio: result.nomeServizio,
-        dataEmissione: result.dataEmissione,
-        importo: result.importo,
-        tipologia: result.tipologia,
-        intestatario: {
-          nome: result.nome,
-          indirizzo: result.indirizzo,
-          telefono: result.telefono,
-          email: result.email,
-          iban: result.iban,
-          pec: result.pec
+      if (result) {
+        const fattura: Fattura = {
+          nomeServizio: result.nomeServizio,
+          dataEmissione: result.dataEmissione,
+          importo: result.importo,
+          tipologia: result.tipologia,
+          intestatario: {
+            nome: result.nome,
+            indirizzo: result.indirizzo,
+            telefono: result.telefono,
+            email: result.email,
+            iban: result.iban,
+            pec: result.pec
+          }
         }
+        this.postFattura(fattura);
       }
-      this.postFattura(fattura);
     });
   }
 
@@ -55,23 +58,25 @@ export class AppComponent implements OnInit {
     });
 
     dialogRef.afterClosed().subscribe(result => {
-      const fattura: Fattura = {
-        _id: result._id,
-        nomeServizio: result.nomeServizio,
-        dataEmissione: result.dataEmissione,
-        importo: result.importo,
-        tipologia: result.tipologia,
-        pagata: result.pagata,
-        intestatario: {
-          nome: result.nome,
-          indirizzo: result.indirizzo,
-          telefono: result.telefono,
-          email: result.email,
-          iban: result.iban,
-          pec: result.pec
+      if (result) {
+        const fattura: Fattura = {
+          _id: result._id,
+          nomeServizio: result.nomeServizio,
+          dataEmissione: result.dataEmissione,
+          importo: result.importo,
+          tipologia: result.tipologia,
+          pagata: result.pagata,
+          intestatario: {
+            nome: result.nome,
+            indirizzo: result.indirizzo,
+            telefono: result.telefono,
+            email: result.email,
+            iban: result.iban,
+            pec: result.pec
+          }
         }
+        this.putFattura(fattura);
       }
-      this.putFattura(fattura);
     });
   }
 
@@ -96,8 +101,20 @@ export class AppComponent implements OnInit {
   }
 
   deleteFattura(fattura: Fattura) {
-    this.fattureService.deleteFattura(fattura).subscribe( (data: any) => {
-      this.fatture = [...this.fatture.filter(fattura => fattura._id !== data._id)];
-    })
+    const dialogRef = this.dialogService.open(ConfirmDialogComponent, {
+      width: '800px',
+      data: {
+        titolo: 'Conferma eliminazione fattura',
+        testo: 'Sei sicuro di voler eliminare la fattura \"' + fattura.nomeServizio + '\"? L\'azione non sarà reversibile.'
+      }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.fattureService.deleteFattura(fattura).subscribe( (data: any) => {
+          this.fatture = [...this.fatture.filter(fattura => fattura._id !== data._id)];
+        })
+      }
+    });
   }
 }
